加入UpdateEdgeCost

+++ d_star_lite_cpp/D_start_Lite/inc/GeneratePath.h
@@ -7,5 +7,5 @@
 /* Matrix Declarations */
 
 /* Function Declarations */
-extern MatrixXd GeneratePath(void);
+extern MatrixXi GeneratePath(void);
 #endif
