+++ d_star_lite_cpp/D_start_Lite/inc/rhs_cal.h
@@ -8,6 +8,6 @@
 /* Matrix Declarations */
 
 /* Function Declarations */
-extern void rhs_cal(RowVectorXd s);
+extern void rhs_cal(RowVectorXi s);
 
 #endif
